+++ src/stactools/core/__init__.py
@@ -29,4 +29,4 @@ __all__ = [
     "move_all_assets",
     "use_fsspec",
 ]
-__version__ = "0.5.2"
+__version__ = "0.5.3"