+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Make computation of statistics and histogram optional for `core.add_raster.add_raster_to_item` ([#466](https://github.com/stac-utils/stactools/issues/466))
+- Make computation of statistics and histogram optional for `core.add_raster.add_raster_to_item` ([#467](https://github.com/stac-utils/stactools/pull/467))
 
 ## [0.5.2] - 2023-09-20