+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.3] - 2023-11-21
+
 ### Changed
 
 - Make computation of statistics and histogram optional for `core.add_raster.add_raster_to_item` ([#467](https://github.com/stac-utils/stactools/pull/467))
@@ -441,7 +443,8 @@ See [#9](https://github.com/stac-utils/stactools/pull/9)
 - `stac.cli.command.layout` for modifying the layout of STACs
 - `stac.browse` for launching a local instance of stac-browser using docker.
 
-[Unreleased]: <https://github.com/stac-utils/stactools/compare/v0.5.2..main>
+[Unreleased]: <https://github.com/stac-utils/stactools/compare/v0.5.3..main>
+[0.5.3]: <https://github.com/stac-utils/stactools/compare/v0.5.2..v0.5.3>
 [0.5.2]: <https://github.com/stac-utils/stactools/compare/v0.5.1..v0.5.2>
 [0.5.1]: <https://github.com/stac-utils/stactools/compare/v0.5.0..v0.5.1>
 [0.5.0]: <https://github.com/stac-utils/stactools/compare/v0.4.8..v0.5.0>